refactors jest test environment setup
