+++ server/server.js
@@ -75,12 +75,8 @@ app.use((err, req, res) => {
 
 app.use(errorHandler);
 
-if (process.env.TEST_ENV == "true") {
-  const server = app.listen(port, () => {
-    console.log(`Test Server running on port ${port}`);
-  });
-  module.exports = { app, server };
-} else {
-  app.listen(port, () => console.log(`Server running on port ${port}`));
-  module.exports = app;
+if (process.env.TEST_ENV !== "true") {
+  app.listen(port, () => console.log(`Server running on port ${port}`)); 
 }
+
+module.exports = app;
